# Conflicts:
#	.github/workflows/ci.yml
#	.github/workflows/codeql.yml
#	.github/workflows/fluent_linter.yml
#	.github/workflows/font_tests.yml
#	.github/workflows/lint.yml
#	.github/workflows/publish_release.yml
#	.github/workflows/publish_website.yml
#	.github/workflows/types_tests.yml
#	test/unit/evaluator_spec.js
#	web/app.js
#	web/pdf_page_view.js
#	web/pdf_print_service.js
#	web/pdf_viewer.js
#	web/viewer.js

+++ src/display/editor/toolbar.js
@@ -48,11 +48,14 @@ class EditorToolbar {
     const editToolbar = (this.#toolbar = document.createElement("div"));
     editToolbar.classList.add("editToolbar", "hidden");
     editToolbar.setAttribute("role", "toolbar");
+
     const signal = this.#editor._uiManager._signal;
-    editToolbar.addEventListener("contextmenu", noContextMenu, { signal });
-    editToolbar.addEventListener("pointerdown", EditorToolbar.#pointerDown, {
-      signal,
-    });
+    if (signal instanceof AbortSignal && !signal.aborted) {
+      editToolbar.addEventListener("contextmenu", noContextMenu, { signal });
+      editToolbar.addEventListener("pointerdown", EditorToolbar.#pointerDown, {
+        signal,
+      });
+    }
 
     const buttons = (this.#buttons = document.createElement("div"));
     buttons.className = "buttons";
@@ -97,6 +100,9 @@ class EditorToolbar {
     // the mouse, we don't want to trigger any focus events on
     // the editor.
     const signal = this.#editor._uiManager._signal;
+    if (!(signal instanceof AbortSignal) || signal.aborted) {
+      return false;
+    }
     element.addEventListener("focusin", this.#focusIn.bind(this), {
       capture: true,
       signal,
@@ -106,6 +112,7 @@ class EditorToolbar {
       signal,
     });
     element.addEventListener("contextmenu", noContextMenu, { signal });
+    return true;
   }
 
   hide() {
@@ -123,17 +130,18 @@ class EditorToolbar {
     const { editorType, _uiManager } = this.#editor;
 
     const button = document.createElement("button");
-    button.className = "delete";
+    button.classList.add("basic", "deleteButton");
     button.tabIndex = 0;
     button.setAttribute("data-l10n-id", EditorToolbar.#l10nRemove[editorType]);
-    this.#addListenersToElement(button);
-    button.addEventListener(
-      "click",
-      e => {
-        _uiManager.delete();
-      },
-      { signal: _uiManager._signal }
-    );
+    if (this.#addListenersToElement(button)) {
+      button.addEventListener(
+        "click",
+        e => {
+          _uiManager.delete();
+        },
+        { signal: _uiManager._signal }
+      );
+    }
     this.#buttons.append(button);
   }
 
@@ -159,7 +167,7 @@ class EditorToolbar {
       return;
     }
     this.#addListenersToElement(button);
-    this.#buttons.prepend(button, this.#divider);
+    this.#buttons.append(button, this.#divider);
     this.#comment = comment;
     comment.toolbar = this;
   }
@@ -214,7 +222,7 @@ class EditorToolbar {
   }
 }
 
-class HighlightToolbar {
+class FloatingToolbar {
   #buttons = null;
 
   #toolbar = null;
@@ -229,15 +237,37 @@ class HighlightToolbar {
     const editToolbar = (this.#toolbar = document.createElement("div"));
     editToolbar.className = "editToolbar";
     editToolbar.setAttribute("role", "toolbar");
-    editToolbar.addEventListener("contextmenu", noContextMenu, {
-      signal: this.#uiManager._signal,
-    });
+
+    const signal = this.#uiManager._signal;
+    if (signal instanceof AbortSignal && !signal.aborted) {
+      editToolbar.addEventListener("contextmenu", noContextMenu, {
+        signal,
+      });
+    }
 
     const buttons = (this.#buttons = document.createElement("div"));
     buttons.className = "buttons";
     editToolbar.append(buttons);
 
-    this.#addHighlightButton();
+    if (this.#uiManager.hasCommentManager()) {
+      this.#makeButton(
+        "commentButton",
+        `pdfjs-comment-floating-button`,
+        "pdfjs-comment-floating-button-label",
+        () => {
+          this.#uiManager.commentSelection("floating_button");
+        }
+      );
+    }
+
+    this.#makeButton(
+      "highlightButton",
+      `pdfjs-highlight-floating-button1`,
+      "pdfjs-highlight-floating-button-label",
+      () => {
+        this.#uiManager.highlightSelection("floating_button");
+      }
+    );
 
     return editToolbar;
   }
@@ -279,26 +309,22 @@ class HighlightToolbar {
     this.#toolbar.remove();
   }
 
-  #addHighlightButton() {
+  #makeButton(buttonClass, l10nId, labelL10nId, clickHandler) {
     const button = document.createElement("button");
-    button.className = "highlightButton";
+    button.classList.add("basic", buttonClass);
     button.tabIndex = 0;
-    button.setAttribute("data-l10n-id", `pdfjs-highlight-floating-button1`);
+    button.setAttribute("data-l10n-id", l10nId);
     const span = document.createElement("span");
     button.append(span);
     span.className = "visuallyHidden";
-    span.setAttribute("data-l10n-id", "pdfjs-highlight-floating-button-label");
+    span.setAttribute("data-l10n-id", labelL10nId);
     const signal = this.#uiManager._signal;
-    button.addEventListener("contextmenu", noContextMenu, { signal });
-    button.addEventListener(
-      "click",
-      () => {
-        this.#uiManager.highlightSelection("floating_button");
-      },
-      { signal }
-    );
+    if (signal instanceof AbortSignal && !signal.aborted) {
+      button.addEventListener("contextmenu", noContextMenu, { signal });
+      button.addEventListener("click", clickHandler, { signal });
+    }
     this.#buttons.append(button);
   }
 }
 
-export { EditorToolbar, HighlightToolbar };
+export { EditorToolbar, FloatingToolbar };
